+++ lib/client.js
@@ -92,6 +92,11 @@ merge(proto, require('./object'));
  */
 merge(proto, require('./bucket'));
 /**
+ * Multipart operations
+ */
+merge(proto, require('./multipart'));
+
+/**
  * ImageClient class
  */
 Client.ImageClient = require('./image')(Client);
