+++ lib/object.js
@@ -20,7 +20,6 @@ var is = require('is-type-of');
 var destroy = require('destroy');
 var eoe = require('end-or-error');
 var urlutil = require('url');
-var util = require('util');
 
 var proto = exports;
 
@@ -42,7 +41,7 @@ proto.put = function* (name, file, options) {
 proto._putData = function* (name, data, options) {
   options = options || {};
   options.headers = options.headers || {};
-  convertMetaToHeaders(options.meta, options.headers);
+  this._convertMetaToHeaders(options.meta, options.headers);
 
   if (!options.headers['Content-Length']) {
     if (data.size === null) {
@@ -70,7 +69,7 @@ proto.putStream = function* (name, stream, options) {
   options = options || {};
   options.headers = options.headers || {};
   options.headers['Transfer-Encoding'] = 'chunked';
-  convertMetaToHeaders(options.meta, options.headers);
+  this._convertMetaToHeaders(options.meta, options.headers);
   var params = this._objectRequestParams('PUT', name, options);
   params.mime = options.mime;
 
@@ -143,7 +142,7 @@ proto.get = function* (name, file, options) {
       writeStream.destroy();
       // should delete the exists file before throw error
       debug('get error: %s, delete the exists file %s', err, file);
-      yield deleteFileSafe(file);
+      yield this._deleteFileSafe(file);
     }
     throw err;
   }
@@ -230,7 +229,7 @@ proto.copy = function* (name, sourceName, options) {
   if (options.meta) {
     options.headers['x-oss-metadata-directive'] = 'REPLACE';
   }
-  convertMetaToHeaders(options.meta, options.headers);
+  this._convertMetaToHeaders(options.meta, options.headers);
 
   if (sourceName[0] !== '/') {
     // copy same bucket object
@@ -341,231 +340,8 @@ proto.signatureUrl = function (name) {
   return url.format();
 };
 
-/**
- * Upload an file to OSS using multipart uploads
- * @param {String} name
- * @param {String|File} file
- * @param {Object} options
- */
-proto.multipartUpload = function* (name, file, options) {
-  var minPartSize = 100 * 1024;
-
-  var fileSize = yield* this._getFileSize(file);
-  if (fileSize < minPartSize) {
-    var data = {
-      stream: this._createStream(file, 0, fileSize),
-      size: fileSize
-    };
-    var ret = yield* this._putData(name, data, options);
-    if (options && options.progress) {
-      yield* options.progress(1);
-    }
-
-    return ret;
-  }
-
-  options = options || {};
-  options.headers = options.headers || {};
-  convertMetaToHeaders(options.meta, options.headers);
-
-  if (options.partSize && options.partSize < minPartSize) {
-    throw new Error('partSize must not be smaller than ' + minPartSize);
-  }
-
-  var result = yield* this._initMultipartUpload(name, options);
-  var uploadId = result.uploadId;
-
-  var partSize = this._getPartSize(fileSize, options.partSize);
-  var partDatas = this._divideParts(file, fileSize, partSize);
-  var numParts = partDatas.length;
-  var parts = [];
-  for (var i = 0; i < numParts; i++) {
-    var partNo = i + 1;
-    var result = yield* this._uploadPart(name, uploadId, partNo, partDatas[i]);
-    parts.push({
-      number: partNo,
-      etag: result.res.headers.etag
-    });
-
-    if (options && options.progress) {
-      yield* options.progress(partNo / numParts);
-    }
-  }
-
-  return yield* this._completeMultipartUpload(name, uploadId, parts, options);
-};
-
-/**
- * List the on-going multipart uploads
- * @param {Object} options
- * @return {Array} the multipart uploads
- */
-proto.listUploads = function* (query, options) {
-  options = options || {};
-  options.subres = 'uploads';
-  var params = this._objectRequestParams('GET', '', options)
-  params.query = query;
-  params.xmlResponse = true;
-  params.successStatuses = [200];
-
-  var result = yield this.request(params);
-  var uploads = result.data.Upload || [];
-  if (!Array.isArray(uploads)) {
-    uploads = [uploads];
-  }
-  uploads = uploads.map(function (up) {
-    return {
-      name: up.Key,
-      uploadId: up.UploadId,
-      initiated: up.Initiated
-    };
-  });
-
-  return {
-    res: result.res,
-    uploads: uploads,
-    bucket: result.data.Bucket,
-    nextKeyMarker: result.data.NextKeyMarker,
-    nextUploadIdMarker: result.data.NextUploadIdMarker,
-    isTruncated: result.data.IsTruncated === 'true'
-  };
-};
-
-/**
- * Abort a multipart upload transaction
- * @param {String} name the object name
- * @param {String} uploadId the upload id
- * @param {Object} options
- */
-proto.abortMultipartUpload = function* (name, uploadId, options) {
-  options = options || {};
-  options.subres = {uploadId: uploadId};
-  var params = this._objectRequestParams('DELETE', name, options);
-  params.successStatuses = [204];
-
-  var result = yield* this.request(params);
-
-  return {
-    res: result.res
-  };
-};
-
-/**
- * Initiate a multipart upload transaction
- * @param {String} name the object name
- * @param {Object} options
- * @return {String} upload id
- */
-proto._initMultipartUpload = function* (name, options) {
-  options = options || {};
-  options.headers = options.headers || {};
-  convertMetaToHeaders(options.meta, options.headers);
-
-  options.subres = 'uploads';
-  var params = this._objectRequestParams('POST', name, options);
-  params.mime = options.mime;
-  params.xmlResponse = true;
-  params.successStatuses = [200];
-
-  var result = yield* this.request(params);
-
-  return {
-    res: result.res,
-    bucket: result.data.Bucket,
-    name: result.data.Key,
-    uploadId: result.data.UploadId
-  };
-};
-
-/**
- * Upload a part in a multipart upload transaction
- * @param {String} name the object name
- * @param {String} uploadId the upload id
- * @param {Integer} partNo the part number
- * @param {Object} data the body data
- * @param {Object} options
- */
-proto._uploadPart = function* (name, uploadId, partNo, data) {
-  var options = {};
-  options.headers = {
-    'Content-Length': data.size
-  };
-
-  options.subres = {
-    partNumber: partNo,
-    uploadId: uploadId
-  };
-  var params = this._objectRequestParams('PUT', name, options);
-  params.mime = options.mime;
-  params.stream = data.stream;
-  params.successStatuses = [200];
-
-  var result = yield* this.request(params);
-
-  return {
-    name: name,
-    etag: result.res.headers.etag,
-    res: result.res
-  };
-};
-
-/**
- * Complete a multipart upload transaction
- * @param {String} name the object name
- * @param {String} uploadId the upload id
- * @param {Array} parts the uploaded parts
- * @param {Object} options
- */
-proto._completeMultipartUpload = function* (name, uploadId, parts, options) {
-  var xml = '<?xml version="1.0" encoding="UTF-8"?>\n<CompleteMultipartUpload>\n';
-  for (var i = 0; i < parts.length; i++) {
-    var p = parts[i];
-    xml += '<Part>\n';
-    xml += '<PartNumber>' + p.number + '</PartNumber>\n';
-    xml += '<ETag>' + p.etag + '</ETag>\n';
-    xml += '</Part>\n';
-  }
-  xml += '</CompleteMultipartUpload>';
-
-  options = options || {};
-  options.subres = {uploadId: uploadId};
-  var params = this._objectRequestParams('POST', name, options);
-  params.mime = 'xml';
-  params.content = xml;
-  params.xmlResponse = true;
-  params.successStatuses = [200];
-  var result = yield* this.request(params);
-
-  return {
-    res: result.res,
-    bucket: result.data.Bucket,
-    name: result.data.Key,
-    etag: result.data.ETag
-  };
-};
-
 proto._objectUrl = function (name) {
   return this._getReqUrl({bucket: this.options.bucket, object: name});
-};
-
-is.file = function (file) {
-  return file.constructor.name === 'File';
-};
-
-/**
- * Get file size
- */
-proto._getFileSize = function* (file) {
-  if (is.buffer(file)) {
-    return file.length;
-  } else if (is.file(file)) {
-    return file.size;
-  } if (is.string(file)) {
-    var stat = yield statFile(file);
-    return stat.size;
-  }
-
-  throw new Error('_getFileSize requires Buffer/File/String.');
 };
 
 /**
@@ -605,79 +381,6 @@ proto._getContent = function* (file) {
   return content;
 };
 
-/*
- * Readable stream for Web File
- */
-var Readable = require('stream').Readable;
-
-function WebFileReadStream(file, options) {
-  if (!(this instanceof WebFileReadStream)) {
-    return new WebFileReadStream(file, options);
-  }
-
-  Readable.call(this, options);
-
-  this.file = file;
-  this.reader = null;
-};
-util.inherits(WebFileReadStream, Readable);
-
-WebFileReadStream.prototype._read = function () {
-  if (!this.reader) {
-    var that = this;
-    that.reader = new FileReader();
-    that.reader.onload = function (e) {
-      that.push(new Buffer(e.target.result, 'binary'));
-      that.push(null);
-    };
-    that.reader.readAsBinaryString(that.file);
-  }
-};
-
-proto._createStream = function (file, start, end) {
-  if (is.file(file)) {
-    return new WebFileReadStream(file.slice(start, end));
-  } else if (is.string(file)) {
-    return fs.createReadStream(file, {
-      start: start,
-      end: end - 1
-    });
-  } else {
-    throw new Error('_createStream requires File/String.');
-  }
-};
-
-proto._getPartSize = function (fileSize, partSize) {
-  var maxNumParts = 10 * 1000;
-  var defaultPartSize = 1 * 1024 * 1024;
-
-  if (!partSize) {
-    return defaultPartSize;
-  }
-
-  return Math.max(
-    Math.ceil(fileSize / maxNumParts),
-    partSize
-  );
-};
-
-proto._divideParts = function (file, fileSize, partSize) {
-  var numParts = Math.ceil(fileSize / partSize);
-
-  var partDatas = [];
-  for (var i = 0; i < numParts; i++) {
-    var start = partSize * i;
-    var end = Math.min(start + partSize, fileSize);
-
-    partDatas.push({
-      stream: this._createStream(file, start, end),
-      size: end - start
-    });
-  }
-
-  return partDatas;
-};
-
 /**
  * generator request params
  * @return {Object} params
@@ -710,7 +413,13 @@ proto._objectName = function (name) {
   return name.replace(/^\/+/, '');
 };
 
-function convertMetaToHeaders(meta, headers) {
+proto._statFile = function (filepath) {
+  return function (callback) {
+    fs.stat(filepath, callback);
+  };
+};
+
+proto._convertMetaToHeaders = function (meta, headers) {
   if (!meta) {
     return;
   }
@@ -718,15 +427,9 @@ function convertMetaToHeaders(meta, headers) {
   for (var k in meta) {
     headers['x-oss-meta-' + k] = meta[k];
   }
-}
+};
 
-function statFile(filepath) {
-  return function (callback) {
-    fs.stat(filepath, callback);
-  };
-}
-
-function deleteFileSafe(filepath) {
+proto._deleteFileSafe = function (filepath) {
   return function (callback) {
     fs.exists(filepath, function (exists) {
       if (!exists) {
@@ -740,4 +443,4 @@ function deleteFileSafe(filepath) {
       });
     });
   };
-}
+};
